Create Incubator software
Coding for and incubator I used a uno R3, a small TFT LCD, a dht22, a nickel plated rain drop sensor, a geared egg turner motor, and  solid state relays. Fan uses a 40 amp relay and a 120 vac computer fan with one leg being tied into the output of the 40 amp relay. the lights are 100 amp and the relay is a 100 amp Solid state relay tied in just like the fan.The water pump is simply a 3 vdc waterpump found on amazon.